AE-2858 AltitudeVisualizationWithGroundOffset uses airport elevation
In X-Plane, the airport elevation determines the elevation of the whole scenary.
In order to have the plane sit on the ground, it needs to be staying at the
airport elevation, and not the runway elevation.

In Google Earth visualization, it might cause small issues when the aircraft
is airborne but close to the ground.