AE-2756 Better detect "flat" time characterisitcs in HDF files
With some recorders, the time is constant and assumed to be the time when
the recorder was started. Sometimes the time parameters jump during the flight,
maybe due to a short power loss. This is often accompanied with masked values
from other parameters.

We will now categorize an HDF has having a constant time as long as its
Minute parameter is constant on average for at least 5 minutes.