AE-2914 Add new KPV Gear Down Selected Duration After Go Around
This new KPV will measure the time it took to raise the Gear after a Go Around.
It will scan for Gear Up Selection up until the following Top Of Climb. If no
Gear Up was obserbed, it will return the total duration until reaching the Top
Of Climb. This will allow to detect go arounds where the gear was not retracted.