AE-2932 ApproachInformation recognizes airports for helicopters
Frames with precise positioning had the restriction that the heading during approach
should match an airport runway. This is not the case for helicopters. For them we just
check the airport distance and pick the shortest one.

In case the helicopter would actually approach from a RWY, the algorithm would still try
to determine it as for any aircrafts.